Move OCP.AppCOnfig, OCP.Comments and OCP.WhatsNew to the server bundle
