+++ core/js/core.json
@@ -14,10 +14,7 @@
 		"sharedialogresharerinfoview.js",
 		"sharedialogshareelistview.js",
 		"contactsmenu_templates.js",
-		"public/appconfig.js",
-		"public/comments.js",
 		"public/publicpage.js",
-		"public/whatsnew.js",
 		"multiselect.js",
 		"oc-requesttoken.js",
 		"setupchecks.js",
