+++ core/src/OCP/index.js
@@ -1,11 +1,17 @@
 /**
  *
  */
-import loader from './loader'
+import * as AppConfig from './appconfig'
+import * as Comments from './comments'
 import initialState from './initialstate'
+import Loader from './loader'
+import * as WhatsNew from './whatsnew'
 
 /** @namespace OCP */
 export default {
-	Loader: loader,
+	AppConfig,
+	Comments,
 	InitialState: initialState,
+	Loader,
+	WhatsNew,
 };
